requirements.txt
# Core Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# LangGraph & LangChain
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-community>=0.0.25

# AI Models
openai==1.12.0
anthropic==0.18.1
langchain-google-genai>=2.1.0

# Web Scraping & Search
crawl4ai>=0.8.0
playwright>=1.49.0
beautifulsoup4>=4.12.0
firecrawl-py>=0.0.5
httpx>=0.26.0
requests>=2.31.0

# Search APIs
google-search-results==2.4.2  # SerpAPI
tavily-python==0.3.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Utilities
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
validators==0.22.0
tldextract==5.1.1

# Data Processing (Optional - commented out to avoid compilation issues)
# pandas>=2.0.0
# numpy>=1.24.0

# Future: Voice & Communication
# elevenlabs==0.2.26
# twilio==8.11.1
